Moved endpoints.js back out of helpers subfolder and into main folder, since having a helpers folder between the apis subfolders was confusing.
