+++ app/lib/api/apis/token/post.js
@@ -5,7 +5,7 @@
 
 import * as m from '../../model';
 import ApiRequest from '../../ApiRequest';
-import { TOKEN_ENDPOINT } from '../helpers/endpoints';
+import { TOKEN_ENDPOINT } from '../endpoints';
 
 const post = (email: string, password: string): Promise<m.ApiResponseData> => {
   const request = new ApiRequest();
