+++ app/lib/api/apis/topics/get.js
@@ -7,7 +7,7 @@ import type { Identifier } from 'types/model';
 
 import * as m from '../../model';
 import ApiRequest from '../../ApiRequest';
-import { TOPICS_ENDPOINT } from '../helpers/endpoints';
+import { TOPICS_ENDPOINT } from '../endpoints';
 
 const get = async (id: Identifier): Promise<m.ApiResponseData> => {
   const request = new ApiRequest();
