+++ app/lib/api/apis/topics/patchContent.js
@@ -8,7 +8,7 @@ import contentItems from 'modules/contentItems';
 
 import * as m from '../../model';
 import ApiRequest from '../../ApiRequest';
-import { TOPICS_ENDPOINT, TOPICS_CONTENT_ENDPOINT } from '../helpers/endpoints';
+import { TOPICS_ENDPOINT, TOPICS_CONTENT_ENDPOINT } from '../endpoints';
 
 const patchContent = (
   topicId: Identifier,
